Mathematical Functions General Project
This program aims to demonstrate an exercise in guessing a number within a given range, usually between 1 and 100, and to deduce during program execution whether the guessed number is lower or higher than the hidden number. The number of attempts is counted, and each attempt is saved in a list. If the player gives up after guessing a number outside the established range, the results obtained with the number to be guessed are displayed on the screen.